requirements.txt
pytest
spacy
sentence-transformers
faiss-cpu
scikit-learn
openai
dotenv
mistralai